[#33] refactor: SendMessage로 코드 이동
