+++ src/libs/api.js
@@ -1,10 +1,10 @@
 import axios from 'axios';
 import { API_BASE_URL } from 'constants/url';
 
-const URL = axios.create({ baseURL: API_BASE_URL });
+const instance = axios.create({ baseURL: API_BASE_URL });
 
 async function getProfileImg() {
-  const response = await URL.get('/profile-images/');
+  const response = await instance.get('/profile-images/');
   const result = response.data;
   const { imageUrls } = result;
   return imageUrls;
@@ -19,7 +19,7 @@ async function postMessage(
   selectFont
 ) {
   try {
-    await URL.post(`/1-10/recipients/${id}/messages/`, {
+    await instance.post(`/1-10/recipients/${id}/messages/`, {
       sender: inputName,
       profileImageURL: profileImg,
       relationship: relationShip,
